Fix kitchen and serverspec so they can be run  * move serverspec to helpers directory and related requires  * update assertions to use pkgutil full path
