+++ .kitchen.yml
@@ -18,10 +18,10 @@ platforms:
   - name: solaris-10.11
     run_list: pkgutil::opencsw
     driver:
-      box: chef/solaris-10.11 # private
+      box: chef/solaris-10.11
 
 suites:
-  - name: pkgutil_package_install
+  - name: pkgutil_add
     run_list: pkgutil_package::install
-  - name: pkgutil_package_remove
+  - name: pkgutil_remove
     run_list: pkgutil_package::remove
